refactoring: 유저관리자 컨트롤러 생성
